# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigVillage Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  gigId: number;
  gigTitle: string;
  providerId: number;
  providerName: string;
  customerUserId?: number | null;
  customerName: string;
  customerContact: string;
  scheduledDate: string;
  status: BookingStatus;
  totalAmount: number;
  notes?: string;
  createdAt: string;
}