# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigVillage Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  gigId: number;
  /** Optional id of the logged-in user placing the booking */
  customerUserId?: number;
  customerName: string;
  customerContact: string;
  scheduledDate: string;
  notes?: string;
}