# lib/api-zod/src/generated/types/createProviderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigVillage Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProviderBody {
  /** Link this provider profile to a user account */
  userId?: number;
  name: string;
  bio: string;
  location: string;
  skills: string[];
}