# lib/api-zod/src/generated/types/provider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigVillage Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Provider {
  id: number;
  userId?: number | null;
  name: string;
  bio: string;
  location: string;
  skills: string[];
  avatar?: string;
  rating?: number;
  reviewCount?: number;
  totalEarnings?: number;
  completedGigs?: number;
  joinedAt: string;
  isVerified: boolean;
}