# lib/api-zod/src/generated/types/signInUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigVillage Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SignInUserBody {
  name: string;
  email: string;
}